todo_rawsql_drf/requirements.txt
Django>=4.2
djangorestframework>=3.15
pytest>=7.0
pytest-django>=4.5
requests>=2.0
